Moved UPNPDev definition to separate header.
